removed common class
