+++ clients/src/main/java/org/apache/kafka/clients/admin/DeleteConsumerGroupsResult.java
@@ -14,20 +14,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.apache.kafka.clients.admin;
 
-import org.apache.kafka.clients.admin.internals.DeleteGroupsResult;
 import org.apache.kafka.common.KafkaFuture;
 
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 
 /**
- * The result of the {@link Admin#deleteConsumerGroups(Collection <String>, DeleteConsumerGroupsOptions)} call.
+ * The result of the {@link Admin#deleteConsumerGroups(Collection)} call.
  */
-public class DeleteConsumerGroupsResult extends DeleteGroupsResult {
-    public DeleteConsumerGroupsResult(final Map<String, KafkaFuture<Void>> futures) {
-        super(futures);
+public class DeleteConsumerGroupsResult {
+    private final Map<String, KafkaFuture<Void>> futures;
+
+    DeleteConsumerGroupsResult(final Map<String, KafkaFuture<Void>> futures) {
+        this.futures = futures;
+    }
+
+    /**
+     * Return a map from group id to futures which can be used to check the status of
+     * individual deletions.
+     */
+    public Map<String, KafkaFuture<Void>> deletedGroups() {
+        Map<String, KafkaFuture<Void>> deletedGroups = new HashMap<>(futures.size());
+        deletedGroups.putAll(futures);
+        return deletedGroups;
+    }
+
+    /**
+     * Return a future which succeeds only if all the consumer group deletions succeed.
+     */
+    public KafkaFuture<Void> all() {
+        return KafkaFuture.allOf(futures.values().toArray(new KafkaFuture<?>[0]));
     }
 }
